[1438]: Move some stuff out of the keeper.go file into a new authz.go file.
